honeypot_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncssh==2.14.2
aiosqlite==0.20.0
aiohttp==3.10.5
psutil==6.0.0
python-dotenv==1.0.1
websockets==13.1
pydantic==2.8.2
cryptography==43.0.1
bcrypt==4.2.0
